move libeinfo_new.c to libeinfo.c
