+++ src/libeinfo/libeinfo.c
@@ -15,983 +15,1147 @@
  *    except according to the terms contained in the LICENSE file.
  */
 
-#include <sys/ioctl.h>
 #include <errno.h>
-#include <inttypes.h>
 #include <stdarg.h>
 #include <stdbool.h>
-#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <strings.h>
+#include <strings.h> /* for strcasecmp(3) */
 #include <syslog.h>
-#ifdef HAVE_TERMCAP
-# include <termcap.h>
+
+#ifdef HAVE_CURSES
+#include <curses.h>
+#include <term.h>
+#else
+#error "no curses found"
 #endif
-#include <unistd.h>
 
 #include "einfo.h"
-#include "helpers.h"
 
-/* Incase we cannot work out how many columns from ioctl, supply a default */
-#define DEFAULT_COLS		 80
+#define EINFO_NONNULL		__attribute__((__nonnull__))
+#define EINFO_NONNULL_POS(X)	__attribute__((__nonnull__(X)))
 
-#define OK			"ok"
-#define NOT_OK			"!!"
+#define LASTCMD_ENV	"EINFO_LASTCMD"
+/*
+ * As soon as the last command run is set from within our code, the env var
+ * becomes useless since it's no longer consistent.
+ */
+#define LASTCMD(X)	do {	\
+	last = (X);		\
+	unsetenv(LASTCMD_ENV);	\
+} while (false)
 
-/* Number of spaces for an indent */
-#define INDENT_WIDTH		2
+/* Spaces per indent */
+#define INDENT_WIDTH	2
+/* Max indent level (in spaces) */
+#define INDENT_MAX	40
 
-/* How wide can the indent go? */
-#define INDENT_MAX		40
-
-/* Default colours */
-#define GOOD                    2
-#define WARN                    3
-#define BAD                     1
-#define HILITE                  6
-#define BRACKET                 4
-
-/* We fallback to these escape codes if termcap isn't available
- * like say /usr isn't mounted */
-#define AF "\033[3%dm"
-#define CE "\033[K"
-#define CH "\033[%dC"
-#define MD "\033[1m"
-#define ME "\033[m"
-#define UP "\033[A"
-
-#define _GET_CAP(_d, _c) strlcpy(_d, tgoto(_c, 0, 0), sizeof(_d));
-#define _ASSIGN_CAP(_v) do {						      \
-		_v = p;							      \
-		p += strlcpy(p, tmp, sizeof(ebuffer) - (p - ebuffer)) + 1;    \
-	} while (0)
-
-/* A pointer to a string to prefix to einfo/ewarn/eerror messages */
-static const char *_eprefix = NULL;
-
-/* Buffers and structures to hold the final colours */
-static char ebuffer[100];
-struct ecolor {
-	ECOLOR color;
-	int def;
-	const char *name;
+struct einfo_term {
+	/* Where to actualy write stuff to */
+	FILE *file;
+	/* Terminal used for curses stuff */
+	TERMINAL *term;
+	/* Is the terminal good for curses use */
+	bool is_good;
+	/* Is the terminal ready for use */
+	bool is_set_up;
 };
-static char nullstr = '\0';
 
-static const struct ecolor ecolors[] = {
-	{ ECOLOR_GOOD,    GOOD,    "good"    },
-	{ ECOLOR_WARN,    WARN,    "warn"    },
-	{ ECOLOR_BAD,     BAD,     "bad"     },
-	{ ECOLOR_HILITE,  HILITE,  "hilite"  },
-	{ ECOLOR_BRACKET, BRACKET, "bracket" },
-	{ ECOLOR_NORMAL,  0,       NULL      },
+/* Used to pick the final color for output */
+struct color_map {
+	int good;
+	bool good_bold;
+	int warn;
+	bool warn_bold;
+	int bad;
+	bool bad_bold;
+	int hilite;
+	bool hilite_bold;
+	int bracket;
+	bool bracket_bold;
 };
-static const char *ecolors_str[ARRAY_SIZE(ecolors)];
 
-static char *flush = NULL;
-static char *up = NULL;
-static char *goto_column = NULL;
-
-static const char *term = NULL;
-static bool term_is_cons25 = false;
-
-/* Termcap buffers and pointers
- * Static buffers suck hard, but some termcap implementations require them */
-#ifdef HAVE_TERMCAP
-static char termcapbuf[2048];
-static char tcapbuf[512];
-#else
-/* No curses support, so we hardcode a list of colour capable terms
- * Only terminals without "color" in the name need to be explicitly listed */
-static const char *const color_terms[] = {
-	"Eterm",
-	"ansi",
-	"con132x25",
-	"con132x30",
-	"con132x43",
-	"con132x60",
-	"con80x25",
-	"con80x28",
-	"con80x30",
-	"con80x43",
-	"con80x50",
-	"con80x60",
-	"cons25",
-	"console",
-	"cygwin",
-	"dtterm",
-	"gnome",
-	"konsole",
-	"kterm",
-	"linux",
-	"linux-c",
-	"mlterm",
-	"putty",
-	"rxvt",
-	"rxvt-cygwin",
-	"rxvt-cygwin-native",
-	"rxvt-unicode",
-	"screen",
-	"screen-bce",
-	"screen-w",
-	"screen.linux",
-	"vt100",
-	"vt220",
-	"wsvt25",
-	"xterm",
-	"xterm-debian",
-	NULL
+static struct einfo_term stdout_term = {
+	.file = NULL,
+	.term = NULL,
+	.is_good = false,
+	.is_set_up = false
 };
-#endif
+static struct einfo_term stderr_term = {
+	.file = NULL,
+	.term = NULL,
+	.is_good = false,
+	.is_set_up = false
+};
+static struct einfo_term *einfo_cur_term = NULL;
 
-#ifndef HAVE_STRLCPY
-static size_t
-strlcat(char *dst, const char *src, size_t size)
+static size_t indent_level = 0;
+
+/* Track the last command run */
+static const char *last = NULL;
+/* Prefixed to messages as "prefix| * msg" */
+static const char *prefix = NULL;
+
+/* Set up the terminals, if needed */
+static void _setupterm(void)
 {
-	char *d = dst;
-	const char *s = src;
-	size_t src_n = size;
-	size_t dst_n;
+	/*
+	 * This var is needed for setupterm() to return an error instead of
+	 * exiting with an error
+	 */
+	int err;
 
-	while (src_n-- != 0 && *d != '\0')
-		d++;
-	dst_n = d - dst;
-	src_n = size - dst_n;
-
-	if (src_n == 0)
-		return dst_n + strlen(src);
-
-	while (*s != '\0') {
-		if (src_n != 1) {
-			*d++ = *s;
-			src_n--;
+	if (!stdout_term.is_set_up) {
+		stdout_term.file = stdout;
+		if (setupterm(NULL, fileno(stdout), &err) != ERR) {
+			stdout_term.term = cur_term;
+			stdout_term.is_good = true;
 		}
-		s++;
+		stdout_term.is_set_up = true;
 	}
-	*d = '\0';
-
-	return dst_n + (s - src);
+	if (!stderr_term.is_set_up) {
+		stderr_term.file = stderr;
+		if (setupterm(NULL, fileno(stderr), &err) != ERR) {
+			stderr_term.term = cur_term;
+			stderr_term.is_good = true;
+		}
+		stderr_term.is_set_up = true;
+	}
 }
-#endif
 
-static bool
-yesno(const char *value)
+/*
+ * Clean up the terminals. Technically not strictly needed since the memory is
+ * freed up on program exit, but let's try to do it right anyway ;)
+ */
+__attribute__((__destructor__))
+static void cleanupterm(void)
 {
-	if (!value) {
+	if (stdout_term.is_set_up && stdout_term.term != NULL) {
+		del_curterm(stdout_term.term);
+	}
+	if (stderr_term.is_set_up && stderr_term.term != NULL) {
+		del_curterm(stderr_term.term);
+	}
+}
+
+/*
+ * Combination of setupterm() and set_curterm(). MUST be called at the start of
+ * a function before any fancy curses stuff is done in order to allow an early
+ * exit if needed.
+ *
+ * t: the target terminal
+ *
+ * return: true if the target terminal is ready for fancy stuff
+ *         false otherwise
+ */
+EINFO_NONNULL
+static bool prepare_term(struct einfo_term *t)
+{
+	_setupterm();
+	if (t->is_good) {
+		set_curterm(t->term);
+	}
+	einfo_cur_term = t;
+
+	return t->is_good;
+}
+
+/*
+ * Checks the truthyness of a string.
+ *
+ * yes/y/true/on/1
+ * no/n/false/off/0
+ *
+ * NULL is false + ENOENT
+ * other strings are false + EINVAL
+ */
+static bool yesno(const char *str)
+{
+	/* Ensure a "true" false does not have a bad errno value */
+	errno = 0;
+
+	if (str == NULL) {
 		errno = ENOENT;
 		return false;
 	}
 
-	if (strcasecmp(value, "yes") == 0 ||
-	    strcasecmp(value, "y") == 0 ||
-	    strcasecmp(value, "true") == 0 ||
-	    strcasecmp(value, "on") == 0 ||
-	    strcasecmp(value, "1") == 0)
+	if (strcasecmp(str, "yes") == 0
+	    || strcasecmp(str, "y") == 0
+	    || strcasecmp(str, "true") == 0
+	    || strcasecmp(str, "on") == 0
+	    || strcasecmp(str, "1") == 0) {
 		return true;
+	}
 
-	if (strcasecmp(value, "no") != 0 &&
-	    strcasecmp(value, "n") != 0 &&
-	    strcasecmp(value, "false") != 0 &&
-	    strcasecmp(value, "off") != 0 &&
-	    strcasecmp(value, "0") != 0)
+	if (strcasecmp(str, "no") != 0
+	    && strcasecmp(str, "n") != 0
+	    && strcasecmp(str, "false") != 0
+	    && strcasecmp(str, "off") != 0
+	    && strcasecmp(str, "0") != 0) {
 		errno = EINVAL;
+	}
 
 	return false;
 }
 
-static bool
-noyes(const char *value)
-{
-	int serrno = errno;
-	bool retval;
-
-	errno = 0;
-	retval = yesno(value);
-	if (errno == 0) {
-		retval = !retval;
-		errno = serrno;
-	}
-
-	return retval;
-}
-
-static bool
-is_quiet(void)
+static bool is_quiet(void)
 {
 	return yesno(getenv("EINFO_QUIET"));
 }
-
-static bool
-is_really_quiet(void)
+static bool is_really_quiet(void)
 {
 	return yesno(getenv("EERROR_QUIET"));
 }
-
-static bool
-is_verbose(void)
+static bool is_verbose(void)
 {
-	return yesno(getenv ("EINFO_VERBOSE"));
+	return yesno(getenv("EINFO_VERBOSE"));
 }
 
-/* Fake tgoto call - very crapy, but works for our needs */
-#ifndef HAVE_TERMCAP
-static char *
-tgoto(const char *cap, int col, int line)
+/* putchar(3)-like for use with tputs */
+static int _putc(int c)
 {
-	static char buf[20];
-	char *p, *e, c, dbuf[6];
-	int oncol = 0, which = line, i;
+	return fputc(c, einfo_cur_term->file);
+}
 
-	p = buf;
-	e = p + sizeof(buf);
-	while ((c = *cap++)) {
-		if (c != '%' || ((c = *cap++) == '%')) {
-			*p++ = c;
-			if (p >= e) {
-				errno = E2BIG;
-				return NULL;
-			}
+/*
+ * Like strtoul(3), but with a maximum returned value. Invalid strings are
+ * silently turned into 0.
+ *
+ * str: string to intify (must be checked by caller for NULL)
+ * max: values larger than this get chomped (including overflow)
+ *
+ * return: the value of str as ulong, or max
+ *         set errno to EINVAL on invalid strings
+ *         set errno to ERANGE if max < str <= ULONG_MAX (no overflow)
+ */
+EINFO_NONNULL
+static unsigned long chomp_strtoul(const char *str, unsigned long max)
+{
+	unsigned long ret = 0;
+	/* Mildly gross, but not truly used while uninitialized */
+	char *endptr;
+
+	errno = 0;
+	ret = strtoul(str, &endptr, 0);
+
+	/* No valid digits */
+	if (endptr == str) {
+		ret = 0;
+		errno = EINVAL;
+	}
+	/* String not completely valid */
+	else if (str[0] != '\0' && endptr[0] != '\0') {
+		ret = 0;
+		errno = EINVAL;
+	}
+	/* ulong overflow */
+	else if (errno == ERANGE) {
+		ret = max;
+	}
+	else if (ret > max) {
+		ret = max;
+		errno = ERANGE;
+	}
+
+	return ret;
+}
+
+/*
+ * Translate EINFO_COLOR env var into a set of colors. Assumes defaults unless
+ * otherwise specified.
+ *
+ * The colors are stored in the env var as "name=color;bold:", for example:
+ *
+ *     EINFO_COLOR="good=1;0:warn=6;1:"
+ *
+ * which sets "good" messages to non-bold red and "warn" messages to bold cyan
+ * assuming 1 and 6 correspond with red and cyan in curses.h, respectively.
+ *
+ * env: the env var itself, must be copied because getenv(3) explicitly warns
+ *      about not modifying the string
+ *
+ * return: the (possibly) modified set of colors
+ *
+ * TODO:
+ * 	- support named colors? (green, green-bold)
+ */
+static struct color_map env_colors(const char *env)
+{
+	char *color_env;
+	char *env_saveptr = NULL;
+	char *entry;
+	char *entry_saveptr = NULL;
+	char *name;
+	char *color;
+	unsigned long parsed_color;
+	char *bold;
+	unsigned long parsed_bold;
+	struct color_map colors = {
+		.good		= COLOR_GREEN,
+		.good_bold	= true,
+		.warn		= COLOR_YELLOW,
+		.warn_bold	= true,
+		.bad		= COLOR_RED,
+		.bad_bold	= true,
+		.hilite		= COLOR_CYAN,
+		.hilite_bold	= true,
+		.bracket	= COLOR_BLUE,
+		.bracket_bold	= true,
+	};
+	int xlat_color [] = {
+		/* I do not expect many people to use black tho */
+		COLOR_BLACK,
+		COLOR_RED,
+		COLOR_GREEN,
+		COLOR_YELLOW,
+		COLOR_BLUE,
+		COLOR_MAGENTA,
+		COLOR_CYAN,
+		COLOR_WHITE,
+	};
+
+	/* No env var, "", and "yes" all mean default colors */
+	if (env == NULL || env[0] == '\0' || yesno(env)) {
+		return colors;
+	}
+
+	color_env = strdup(env);
+
+	/* Go through each entry in the env var, skip malformed ones */
+	for (entry = strtok_r(color_env, ":", &env_saveptr);
+	    entry != NULL;
+	    entry = strtok_r(NULL, ":", &env_saveptr)) {
+		/*
+		 * The saveptr used for parsing an entry should be reset, see
+		 * NOTES in strtok_r(3)
+		 */
+		entry_saveptr = NULL;
+
+		name = strtok_r(entry, "=", &entry_saveptr);
+		if (name == NULL) {
 			continue;
 		}
-		switch (c) {
-		case '3':
-		case '2':
-		case 'd':
-			i = 0;
-			do
-				dbuf[i++] = which % 10 | '0';
-			while ((which /= 10));
-			if (c != 'd') {
-				c -= '0';
-				if (i > c) {
-					errno = EINVAL;
-					return NULL;
-				}
-				while (i < c)
-					dbuf[i++] = '0';
-			}
-			if (p + i >= e) {
-				errno = E2BIG;
-				return NULL;
-			}
-			do
-				*p++ = dbuf[--i];
-			while (i);
-			break;
-		case 'r':
-			oncol = 0;
-			break;
-		case 'i':
-			col++;
-			line++;
-			which++;
+		color = strtok_r(NULL, ";", &entry_saveptr);
+		if (color == NULL) {
 			continue;
-		default:
-			errno = EINVAL;
-			return NULL;
+		}
+		bold = strtok_r(NULL, ";", &entry_saveptr);
+		if (bold == NULL) {
+			continue;
+		}
+		/* Only 8 colors are defined by curses */
+		parsed_color = chomp_strtoul(color, 7);
+		if (errno != 0) {
+			continue;
+		}
+		parsed_color = xlat_color[parsed_color];
+		parsed_bold = chomp_strtoul(bold, 1);
+		if (errno != 0) {
+			continue;
 		}
 
-		oncol = 1 - oncol;
-		which = oncol ? col : line;
-	}
-	*p = '\0';
-	return buf;
-}
-#endif
-
-static bool
-colour_terminal(FILE * EINFO_RESTRICT f)
-{
-	static int in_colour = -1;
-	char *e, *ee, *end, *d, *p;
-	int c;
-	const char *_af = NULL, *_ce = NULL, *_ch = NULL;
-	const char *_md = NULL, *_me = NULL, *_up = NULL;
-	const char *bold;
-	char tmp[100];
-	unsigned int i = 0;
-#ifdef HAVE_TERMCAP
-	char *bp;
-#endif
-
-	if (f && !isatty(fileno(f)))
-		return false;
-
-	if (noyes(getenv("EINFO_COLOR")))
-		return false;
-
-	if (in_colour == 0)
-		return false;
-	if (in_colour == 1)
-		return true;
-
-	term_is_cons25 = false;
-	if (!term) {
-		term = getenv("TERM");
-		if (!term)
-			return false;
-	}
-	if (strcmp(term, "cons25") == 0)
-		term_is_cons25 = true;
-
-#ifdef HAVE_TERMCAP
-	/* Check termcap to see if we can do colour or not */
-	if (tgetent(termcapbuf, term) == 1) {
-		bp = tcapbuf;
-		_af = tgetstr("AF", &bp);
-		_ce = tgetstr("ce", &bp);
-		_ch = tgetstr("ch", &bp);
-		/* Our ch use also works with RI .... for now */
-		if (!_ch)
-			_ch = tgetstr("RI", &bp);
-		_md = tgetstr("md", &bp);
-		_me = tgetstr("me", &bp);
-		_up = tgetstr("up", &bp);
-	}
-
-	/* Cheat here as vanilla BSD has the whole termcap info in /usr
-	 * which is not available to us when we boot */
-	if (term_is_cons25 || strcmp(term, "wsvt25") == 0) {
-#else
-		if (strstr(term, "color"))
-			in_colour = 1;
-
-		while (color_terms[i] && in_colour != 1) {
-			if (strcmp(color_terms[i], term) == 0) {
-				in_colour = 1;
-			}
-			i++;
+		if (strcasecmp(name, "good") == 0) {
+			colors.good = parsed_color;
+			colors.good_bold = parsed_bold;
+		} else if (strcasecmp(name, "warn") == 0) {
+			colors.warn = parsed_color;
+			colors.warn_bold = parsed_bold;
+		} else if (strcasecmp(name, "bad") == 0) {
+			colors.bad = parsed_color;
+			colors.bad_bold = parsed_bold;
+		} else if (strcasecmp(name, "hilite") == 0) {
+			colors.hilite = parsed_color;
+			colors.hilite_bold = parsed_bold;
+		} else if (strcasecmp(name, "bracket") == 0) {
+			colors.bracket = parsed_color;
+			colors.bracket_bold = parsed_bold;
 		}
-
-		if (in_colour != 1) {
-			in_colour = 0;
-			return false;
+		/* Just for symmetry */
+		else {
+			continue;
 		}
-#endif
-		if (!_af)
-			_af = AF;
-		if (!_ce)
-			_ce = CE;
-		if (!_ch)
-			_ch = CH;
-		if (!_md)
-			_md = MD;
-		if (!_me)
-			_me = ME;
-		if (!_up)
-			_up = UP;
-#ifdef HAVE_TERMCAP
 	}
 
-	if (!_af || !_ce || !_me || !_md || !_up) {
-		in_colour = 0;
-		return false;
-	}
-
-	/* Many termcap databases don't have ch or RI even though they
-	 * do work */
-	if (!_ch)
-		_ch = CH;
-#endif
-
-	/* Now setup our colours */
-	p = ebuffer;
-	for (i = 0; i < ARRAY_SIZE(ecolors); ++i) {
-		tmp[0] = '\0';
-		if (ecolors[i].name) {
-			bold = _md;
-			c = ecolors[i].def;
-
-			/* See if the user wants to override the colour
-			 * We use a :col;bold: format like 2;1: for bold green
-			 * and 1;0: for a normal red */
-			if ((e = getenv("EINFO_COLOR"))) {
-				ee = strstr(e, ecolors[i].name);
-				if (ee)
-					ee += strlen(ecolors[i].name);
-
-				if (ee && *ee == '=') {
-					d = strdup(ee + 1);
-					if (d) {
-						end = strchr(d, ':');
-						if (end)
-							*end = '\0';
-						c = atoi(d);
-						end = strchr(d, ';');
-						if (end && *++end == '0')
-							bold = _me;
-						free(d);
-					}
-				}
-			}
-			strlcpy(tmp, tgoto(bold, 0, 0), sizeof(tmp));
-			strlcat(tmp, tgoto(_af, 0, c & 0x07), sizeof(tmp));
-		} else
-			_GET_CAP(tmp, _me);
-
-		if (tmp[0])
-			_ASSIGN_CAP(ecolors_str[i]);
-		else
-			ecolors_str[i] = &nullstr;
-	}
-
-	_GET_CAP(tmp, _ce);
-	_ASSIGN_CAP(flush);
-	_GET_CAP(tmp, _up);
-	_ASSIGN_CAP(up);
-	strlcpy(tmp, _ch, sizeof(tmp));
-	_ASSIGN_CAP(goto_column);
-
-	in_colour = 1;
-	return true;
+	free(color_env);
+	return colors;
 }
 
-static int
-get_term_columns(FILE * EINFO_RESTRICT stream)
+/*
+ * Set the output color mode
+ *
+ * t: the terminal to set the color for
+ * color: the color to being using
+ */
+EINFO_NONNULL
+static void _ecolor(struct einfo_term *t, ECOLOR color)
 {
-	struct winsize ws;
-	char *env = getenv("COLUMNS");
-	char *p;
-	int i;
+	const char *env_temp = NULL;
+	struct color_map colors;
+	int target_color;
+	bool target_bold;
+	const char *color_str [] = { NULL, NULL };
 
-	if (env) {
-		i = strtoimax(env, &p, 10);
-		if (!*p)
-			return i;
+	if (!prepare_term(t)) {
+		return;
 	}
 
-	if (ioctl(fileno(stream), TIOCGWINSZ, &ws) == 0)
-		return ws.ws_col;
-
-	return DEFAULT_COLS;
-}
-
-void
-eprefix(const char *EINFO_RESTRICT prefix)
-{
-	_eprefix = prefix;
-}
-
-static void EINFO_PRINTF(2, 0)
-elogv(int level, const char *EINFO_RESTRICT fmt, va_list ap)
-{
-	char *e = getenv("EINFO_LOG");
-	va_list apc;
-
-	if (fmt && e) {
-		closelog();
-		openlog(e, LOG_PID, LOG_DAEMON);
-		va_copy(apc, ap);
-		vsyslog(level, fmt, apc);
-		va_end(apc);
-		closelog();
+	env_temp = getenv("EINFO_COLOR");
+	if (!yesno(env_temp)) {
+		/* Only skip colors on a true "no" */
+		if (errno == 0) {
+			return;
+		}
 	}
+	colors = env_colors(env_temp);
+
+	if (color != ECOLOR_NORMAL) {
+		if (color == ECOLOR_GOOD) {
+			target_bold = colors.good_bold;
+			target_color = colors.good;
+		} else if (color == ECOLOR_WARN) {
+			target_bold = colors.warn_bold;
+			target_color = colors.warn;
+		} else if (color == ECOLOR_BAD) {
+			target_bold = colors.bad_bold;
+			target_color = colors.bad;
+		} else if (color == ECOLOR_HILITE) {
+			target_bold = colors.hilite_bold;
+			target_color = colors.hilite;
+		} else if (color == ECOLOR_BRACKET) {
+			target_bold = colors.bracket_bold;
+			target_color = colors.bracket;
+		}
+		color_str[0] = target_bold
+		    ? enter_bold_mode
+		    : exit_attribute_mode;
+		color_str[1] = tiparm(set_a_foreground, target_color);
+	}
+	/* ECOLOR_NORMAL */
+	else {
+		color_str[0] = tiparm(orig_pair);
+		color_str[1] = tiparm(exit_attribute_mode);
+	}
+
+	tputs(color_str[0], 1, _putc);
+	tputs(color_str[1], 1, _putc);
 }
 
-void
-elog(int level, const char *EINFO_RESTRICT fmt, ...)
+/*
+ * Move up one row
+ *
+ * t: terminal to move in
+ */
+EINFO_NONNULL
+static void move_up(struct einfo_term *t)
 {
+	const char *move_str = "";
+
+	if (!prepare_term(t)) {
+		return;
+	}
+
+	move_str = tiparm(cursor_up);
+	tputs(move_str, 1, _putc);
+}
+
+/*
+ * Move to a specified column
+ *
+ * t: terminal to move in
+ * col: column to move to
+ *      >0 is counting from the left edge
+ *      <0 is counting from the right edge
+ */
+EINFO_NONNULL
+static void move_col(struct einfo_term *t, int col)
+{
+	const char *move_str = "";
+
+	if (!prepare_term(t)) {
+		return;
+	}
+
+	if (col == 0) {
+		return;
+	} else if (col < 0) {
+		col += columns;
+	}
+
+	move_str = tiparm(column_address, col);
+	tputs(move_str, 1, _putc);
+}
+
+/*
+ * Perform the indent, if needed
+ *
+ * f: the file that is indented into
+ *
+ * return: like fprintf(3)
+ */
+EINFO_NONNULL
+static int _eindent(FILE *f)
+{
+	/* Pre-terminated :) */
+	char indent [INDENT_MAX + 1] = { 0 };
+	const char *env_level_str = getenv("EINFO_INDENT");
+	size_t env_level = 0;
+
+	/*
+	 * Use the indent level from the env if it exists and is valid. Given as
+	 * a count of spaces.
+	 */
+	if (env_level_str != NULL) {
+		env_level = chomp_strtoul(env_level_str, INDENT_MAX);
+
+		if (env_level == 0 && errno == 0) {
+			return 0;
+		} else if (errno != EINVAL) {
+			memset(indent, ' ', env_level);
+			return fprintf(f, "%s", indent);
+		}
+	}
+
+	/*
+	 * The indent_level bounds are checked in eindent()/eoutdent(), this is
+	 * just to skip unneeded work.
+	 */
+	if (indent_level > 0) {
+		memset(indent, ' ', indent_level * INDENT_WIDTH);
+		return fprintf(f, "%s", indent);
+	}
+
+	return 0;
+}
+
+/*
+ * Actually print the einfo message
+ *
+ * t: the terminal to print the message to
+ * color: the color to use for the einfo marker
+ * fmt: the printf-style format for the message
+ * va: the args to feed vfprintf(3)
+ *
+ * return: like the printf-family of functions
+ */
+EINFO_NONNULL
+EINFO_PRINTF(3, 0)
+static int _einfo(
+    struct einfo_term *t,
+    ECOLOR color,
+    const char *EINFO_RESTRICT fmt,
+    va_list va)
+{
+	const char *last_env = getenv(LASTCMD_ENV);
+
+	int ret = 0;
 	va_list ap;
 
-	va_start(ap, fmt);
-	elogv(level, fmt, ap);
-	va_end(ap);
-}
+	va_copy(ap, va);
 
-static int
-_eindent(FILE * EINFO_RESTRICT stream)
-{
-	char *env = getenv("EINFO_INDENT");
-	int amount = 0;
-	char indent[INDENT_MAX];
-
-	if (env) {
-		errno = 0;
-		amount = strtoimax(env, NULL, 0);
-		if (errno != 0 || amount < 0)
-			amount = 0;
-		else if (amount > INDENT_MAX)
-			amount = INDENT_MAX;
-
-		if (amount > 0)
-			memset(indent, ' ', (size_t)amount);
+	/* Let's assume the env var is "more correct" */
+	if (last_env != NULL) {
+		last = last_env;
 	}
 
-	/* Terminate it */
-	memset(indent + amount, 0, 1);
-	return fprintf(stream, "%s", indent);
+	/*
+	 * Print a newline on non-curses terms if the last command is a
+	 * "no newline" variant. Why only non-curses? IDFK.
+	 */
+	if (!prepare_term(t)
+	    && last != NULL
+	    && last[strlen(last) - 1] == 'n'
+	    && strcmp(last, "ewarn") != 0) {
+		fprintf(t->file, "\n");
+	}
+
+	if (prefix) {
+		_ecolor(t, color);
+		ret += fprintf(t->file, "%s", prefix);
+		_ecolor(t, ECOLOR_NORMAL);
+		ret += fprintf(t->file, "|");
+	}
+	_ecolor(t, color);
+	ret += fprintf(t->file, " * ");
+	_ecolor(t, ECOLOR_NORMAL);
+	ret += _eindent(t->file);
+	ret += vfprintf(t->file, fmt, ap);
+
+	if (last_env != NULL) {
+		/* The env var has served its purpose */
+		unsetenv(LASTCMD_ENV);
+	}
+	last = NULL;
+
+	va_end(ap);
+	return ret;
 }
 
-static const char *
-_ecolor(FILE * EINFO_RESTRICT f, ECOLOR color)
+/*
+ * Print the first part of the eend if applicable
+ *
+ * args: similar to _einfo()
+ * fmt: potentially NULL or empty
+ *
+ * return: similar to _einfo()
+ */
+EINFO_NONNULL_POS(1)
+EINFO_PRINTF(4, 0)
+static int _eend_message(
+    struct einfo_term *t,
+    int retval,
+    ECOLOR color,
+    const char *EINFO_RESTRICT fmt,
+    va_list va)
 {
-	unsigned int i;
+	int ret;
+	va_list ap;
 
-	if (!colour_terminal(f))
-		return "";
+	/* Nothing to see here */
+	if (retval == 0 || fmt == NULL || fmt[0] == '\0') {
+		return 0;
+	}
 
-	for (i = 0; i < ARRAY_SIZE(ecolors); ++i)
-		if (ecolors[i].color == color)
-			return ecolors_str[i];
+	va_copy(ap, va);
+
+	ret = _einfo(t, color, fmt, ap);
+	ret += fprintf(t->file, "\n");
+
+	va_end(ap);
+	return ret;
+}
+
+/*
+ * Print the status part of eend
+ *
+ * t: the terminal to print to
+ * col: the column where the status begins, measured from the initial ' '
+ *      -1 to align at the right edge of the terminal
+ * color: the color to print with
+ * msg: message to include in the brackets
+ */
+EINFO_NONNULL
+static int _eend_status(
+    struct einfo_term *t,
+    int col,
+    ECOLOR color,
+    const char *EINFO_RESTRICT msg)
+{
+	int ret;
+
+	if (msg[0] == '\0') {
+		msg = "...";
+	}
+
+	if (col == -1) {
+		col = -(strlen(msg) + 5);
+	}
+
+	move_up(t);
+	move_col(t, col);
+	_ecolor(t, ECOLOR_BRACKET);
+	ret = fprintf(t->file, " [ ");
+	_ecolor(t, color);
+	ret += fprintf(t->file, "%s", msg);
+	_ecolor(t, ECOLOR_BRACKET);
+	ret += fprintf(t->file, " ]\n");
+	_ecolor(t, ECOLOR_NORMAL);
+
+	return ret;
+}
+
+/*
+ * Print an einfo message to syslog (without the " * " or terminators). Requires
+ * an ID to be given in EINFO_LOG env var.
+ *
+ * level: log level, see syslog(3)
+ * fmt: the printf-style format for the message
+ * va: the args to feed vsyslog(3)
+ */
+EINFO_NONNULL
+EINFO_PRINTF(2, 0)
+static void _elog(int level, const char *EINFO_RESTRICT fmt, va_list va)
+{
+	const char *id = getenv("EINFO_LOG");
+
+	va_list ap;
+
+	if (id) {
+		va_copy(ap, va);
+
+		openlog(id, LOG_PID, LOG_DAEMON);
+		vsyslog(level, fmt, ap);
+		/* Only Linux seems to mention closelog(3) being optional */
+		closelog();
+
+		va_end(ap);
+	}
+}
+
+const char *ecolor(ECOLOR color)
+{
+	(void)color;
 	return "";
 }
 
-const char *
-ecolor(ECOLOR color)
+void elog(int level, const char * EINFO_RESTRICT fmt, ...)
 {
-	FILE *f = stdout;
+	va_list ap;
 
-	/* Try and guess a valid tty */
-	if (!isatty(fileno(f))) {
-		f = stderr;
-		if (!isatty(fileno(f))) {
-			f = stdin;
-			if (!isatty(fileno(f)))
-				f = NULL;
-		}
+	if (!fmt) {
+		return;
 	}
 
-	return _ecolor(f, color);
+	va_start(ap, fmt);
+
+	_elog(level, fmt, ap);
+
+	va_end(ap);
 }
 
-#define LASTCMD(_cmd) {							      \
-		unsetenv("EINFO_LASTCMD");				      \
-		setenv("EINFO_LASTCMD", _cmd, 1);			      \
+int einfon(const char * EINFO_RESTRICT fmt, ...)
+{
+	int ret;
+	va_list ap;
+
+	if (!fmt || is_quiet()) {
+		return 0;
 	}
 
-static int EINFO_PRINTF(3, 0)
-	_einfo(FILE *f, ECOLOR color, const char *EINFO_RESTRICT fmt, va_list va)
-{
-	int retval = 0;
-	char *last = getenv("EINFO_LASTCMD");
-	va_list ap;
-
-	if (last &&
-	    !colour_terminal(f) &&
-	    strcmp(last, "ewarn") != 0 &&
-	    last[strlen(last) - 1] == 'n')
-		fprintf(f, "\n");
-	if (_eprefix)
-		fprintf(f, "%s%s%s|", _ecolor(f, color), _eprefix, _ecolor(f, ECOLOR_NORMAL));
-	fprintf(f, " %s*%s ", _ecolor(f, color), _ecolor(f, ECOLOR_NORMAL));
-	retval += _eindent(f);
-	va_copy(ap, va);
-	retval += vfprintf(f, fmt, ap) + 3;
-	va_end(ap); \
-	if (colour_terminal(f))
-		fprintf(f, "%s", flush);
-	return retval;
-}
-
-#define _einfovn(fmt, ap) _einfo(stdout, ECOLOR_GOOD, fmt, ap)
-#define _ewarnvn(fmt, ap) _einfo(stderr, ECOLOR_WARN, fmt, ap)
-#define _eerrorvn(fmt, ap) _einfo(stderr, ECOLOR_BAD, fmt, ap)
-
-int
-einfon(const char *EINFO_RESTRICT fmt, ...)
-{
-	int retval;
-	va_list ap;
-
-	if (!fmt || is_quiet())
-		return 0;
 	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
-	va_end(ap);
-	LASTCMD("einfon");
-	return retval;
-}
 
-int
-ewarnn(const char *EINFO_RESTRICT fmt, ...)
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+/* TODO: why is this not sent to syslog? */
+int ewarnn(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_quiet())
+	if (!fmt || is_quiet()) {
 		return 0;
-	va_start(ap, fmt);
-	retval = _ewarnvn(fmt, ap);
-	va_end(ap);
-	LASTCMD("ewarnn");
-	return retval;
-}
+	}
 
-int
-eerrorn(const char *EINFO_RESTRICT fmt, ...)
+	va_start(ap, fmt);
+
+	ret = _einfo(&stderr_term, ECOLOR_WARN, fmt, ap);
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+/* TODO: why is this not sent to syslog? */
+int eerrorn(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_really_quiet())
+	if (!fmt || is_really_quiet()) {
 		return 0;
-	va_start(ap, fmt);
-	retval = _eerrorvn(fmt, ap);
-	va_end(ap);
-	LASTCMD("errorn");
-	return retval;
-}
+	}
 
-int
-einfo(const char *EINFO_RESTRICT fmt, ...)
+	va_start(ap, fmt);
+
+	ret = _einfo(&stderr_term, ECOLOR_BAD, fmt, ap);
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+int einfo(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_quiet())
+	if (!fmt || is_quiet()) {
 		return 0;
-	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
-	retval += printf("\n");
-	va_end(ap);
-	LASTCMD("einfo");
-	return retval;
-}
+	}
 
-int
-ewarn(const char *EINFO_RESTRICT fmt, ...)
+	va_start(ap, fmt);
+
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+	ret += fprintf(stdout_term.file, "\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+int ewarn(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_quiet())
+	if (!fmt || is_quiet()) {
 		return 0;
-	va_start(ap, fmt);
-	elogv(LOG_WARNING, fmt, ap);
-	retval = _ewarnvn(fmt, ap);
-	retval += fprintf(stderr, "\n");
-	va_end(ap);
-	LASTCMD("ewarn");
-	return retval;
-}
+	}
 
-void
-ewarnx(const char *EINFO_RESTRICT fmt, ...)
+	va_start(ap, fmt);
+
+	_elog(LOG_WARNING, fmt, ap);
+	ret = _einfo(&stderr_term, ECOLOR_WARN, fmt, ap);
+	ret += fprintf(stderr_term.file, "\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+void ewarnx(const char * EINFO_RESTRICT fmt, ...)
 {
 	va_list ap;
 
 	if (fmt && !is_quiet()) {
 		va_start(ap, fmt);
-		elogv(LOG_WARNING, fmt, ap);
-		_ewarnvn(fmt, ap);
+
+		_elog(LOG_WARNING, fmt, ap);
+		_einfo(&stderr_term, ECOLOR_WARN, fmt, ap);
+		fprintf(stderr_term.file, "\n");
+
 		va_end(ap);
-		fprintf(stderr, "\n");
 	}
+
 	exit(EXIT_FAILURE);
 }
-
-int
-eerror(const char *EINFO_RESTRICT fmt, ...)
+int eerror(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_really_quiet())
+	if (!fmt || is_really_quiet()) {
 		return 0;
-	va_start(ap, fmt);
-	elogv(LOG_ERR, fmt, ap);
-	retval = _eerrorvn(fmt, ap);
-	va_end(ap);
-	retval += fprintf(stderr, "\n");
-	LASTCMD("eerror");
-	return retval;
-}
+	}
 
-void
-eerrorx(const char *EINFO_RESTRICT fmt, ...)
+	va_start(ap, fmt);
+
+	_elog(LOG_ERR, fmt, ap);
+	ret = _einfo(&stderr_term, ECOLOR_BAD, fmt, ap);
+	ret += fprintf(stderr_term.file, "\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+void eerrorx(const char * EINFO_RESTRICT fmt, ...)
 {
 	va_list ap;
 
 	if (fmt && !is_really_quiet()) {
 		va_start(ap, fmt);
-		elogv(LOG_ERR, fmt, ap);
-		_eerrorvn(fmt, ap);
+
+		_elog(LOG_ERR, fmt, ap);
+		_einfo(&stderr_term, ECOLOR_BAD, fmt, ap);
+		fprintf(stderr_term.file, "\n");
+
 		va_end(ap);
-		fprintf(stderr, "\n");
 	}
+
 	exit(EXIT_FAILURE);
 }
 
-int
-ebegin(const char *EINFO_RESTRICT fmt, ...)
+int einfovn(const char * EINFO_RESTRICT fmt, ...)
 {
-	int retval;
+	int ret;
 	va_list ap;
 
-	if (!fmt || is_quiet())
+	if (!fmt || !is_verbose()) {
 		return 0;
+	}
+
 	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
+
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+
 	va_end(ap);
-	retval += printf(" ...");
-	if (colour_terminal(stdout))
-		retval += printf("\n");
-	LASTCMD("ebegin");
-	return retval;
+	LASTCMD(__func__);
+	return ret;
+}
+/* TODO: why is this not sent to syslog? */
+int ewarnvn(const char * EINFO_RESTRICT fmt, ...)
+{
+	int ret;
+	va_list ap;
+
+	if (!fmt || !is_verbose()) {
+		return 0;
+	}
+
+	va_start(ap, fmt);
+
+	ret = _einfo(&stderr_term, ECOLOR_WARN, fmt, ap);
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+/* XXX: Not implemented originally? */
+int ebeginvn(const char * EINFO_RESTRICT fmt, ...)
+{
+	(void)fmt;
+	return 0;
+}
+int eendvn(int n, const char * EINFO_RESTRICT fmt, ...)
+{
+	(void)n;
+	(void)fmt;
+	return 0;
+}
+int ewendvn(int n, const char * EINFO_RESTRICT fmt, ...)
+{
+	(void)n;
+	(void)fmt;
+	return 0;
+}
+/* XXX: End not implemented */
+int einfov(const char * EINFO_RESTRICT fmt, ...)
+{
+	int ret;
+	va_list ap;
+
+	if (!fmt || !is_verbose()) {
+		return 0;
+	}
+
+	va_start(ap, fmt);
+
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+	ret += fprintf(stdout_term.file, "\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+/* TODO: why is this not sent to syslog? */
+int ewarnv(const char * EINFO_RESTRICT fmt, ...)
+{
+	int ret;
+	va_list ap;
+
+	if (!fmt || !is_verbose()) {
+		return 0;
+	}
+
+	va_start(ap, fmt);
+
+	ret = _einfo(&stderr_term, ECOLOR_WARN, fmt, ap);
+	ret += fprintf(stderr_term.file, "\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
 }
 
-static void
-_eend(FILE * EINFO_RESTRICT fp, int col, ECOLOR color, const char *msg)
+int ebeginv(const char * EINFO_RESTRICT fmt, ...)
 {
-	int i;
-	int cols;
+	int ret;
+	va_list ap;
 
-	if (!msg)
-		return;
-
-	cols = get_term_columns(fp) - (strlen(msg) + 5);
-
-	/* cons25 is special - we need to remove one char, otherwise things
-	 * do not align properly at all. */
-	if (!term) {
-		term = getenv("TERM");
-		if (term && strcmp(term, "cons25") == 0)
-			term_is_cons25 = true;
-		else
-			term_is_cons25 = false;
+	if (!fmt || !is_verbose()) {
+		return 0;
 	}
-	if (term_is_cons25)
-		cols--;
 
-	if (cols > 0 && colour_terminal(fp)) {
-		fprintf(fp, "%s%s %s[%s %s %s]%s\n", up, tgoto(goto_column, 0, cols),
-		    ecolor(ECOLOR_BRACKET), ecolor(color), msg,
-		    ecolor(ECOLOR_BRACKET), ecolor(ECOLOR_NORMAL));
+	va_start(ap, fmt);
+
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+	ret += fprintf(stdout_term.file, " ...\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+int ebegin(const char * EINFO_RESTRICT fmt, ...)
+{
+	int ret;
+	va_list ap;
+
+	if (!fmt || is_quiet()) {
+		return 0;
+	}
+
+	va_start(ap, fmt);
+
+	ret = _einfo(&stdout_term, ECOLOR_GOOD, fmt, ap);
+	ret += fprintf(stdout_term.file, " ...\n");
+
+	va_end(ap);
+	LASTCMD(__func__);
+	return ret;
+}
+
+int eend(int retval, const char * EINFO_RESTRICT fmt, ...)
+{
+	va_list ap;
+	struct einfo_term *term;
+	ECOLOR color;
+	const char *msg;
+
+	if (is_quiet()) {
+		return retval;
+	}
+
+	va_start(ap, fmt);
+
+	if (retval == 0) {
+		term = &stdout_term;
+		color = ECOLOR_GOOD;
+		msg = "ok";
 	} else {
-		if (col > 0)
-			for (i = 0; i < cols - col; i++)
-				fprintf(fp, " ");
-		fprintf(fp, " [ %s ]\n", msg);
+		term = &stderr_term;
+		color = ECOLOR_BAD;
+		msg = "!!";
 	}
-}
+	_eend_message(term, retval, color, fmt, ap);
+	_eend_status(term, -1, color, msg);
 
-static int EINFO_PRINTF(3, 0)
-_do_eend(const char *cmd, int retval,
-    const char *EINFO_RESTRICT fmt, va_list ap)
-{
-	int col = 0;
-	FILE *fp = stdout;
-	va_list apc;
-
-	if (fmt && *fmt != '\0' && retval != 0) {
-		fp = stderr;
-		va_copy(apc, ap);
-		if (strcmp(cmd, "ewend") == 0)
-			col = _ewarnvn(fmt, apc);
-		else
-			col = _eerrorvn(fmt, apc);
-		col += fprintf(fp, "\n");
-		va_end(apc);
-	}
-	_eend(fp, col,
-	    retval == 0 ? ECOLOR_GOOD : ECOLOR_BAD,
-	    retval == 0 ? OK : NOT_OK);
+	va_end(ap);
+	LASTCMD(__func__);
 	return retval;
 }
-
-int
-eend(int retval, const char *EINFO_RESTRICT fmt, ...)
+int ewend(int retval, const char * EINFO_RESTRICT fmt, ...)
 {
 	va_list ap;
+	struct einfo_term *term;
+	ECOLOR color;
+	const char *msg;
 
-	if (is_quiet())
+	if (is_quiet()) {
 		return retval;
+	}
+
 	va_start(ap, fmt);
-	_do_eend("eend", retval, fmt, ap);
+
+	if (retval == 0) {
+		term = &stdout_term;
+		color = ECOLOR_GOOD;
+		msg = "ok";
+	} else {
+		term = &stderr_term;
+		color = ECOLOR_WARN;
+		msg = "!!";
+	}
+	_eend_message(term, retval, color, fmt, ap);
+
+	/* Do not use the warn color for "!!" */
+	if (retval != 0) {
+		color = ECOLOR_BAD;
+	}
+	_eend_status(term, -1, color, msg);
+
 	va_end(ap);
-	LASTCMD("eend");
+	LASTCMD(__func__);
 	return retval;
 }
+void ebracket(int col, ECOLOR color, const char * EINFO_RESTRICT msg)
+{
+	if (msg == NULL) {
+		msg = "";
+	}
 
-int
-ewend(int retval, const char *EINFO_RESTRICT fmt, ...)
+	_eend_status(&stdout_term, col, color, msg);
+
+	LASTCMD(__func__);
+}
+
+int eendv(int retval, const char * EINFO_RESTRICT fmt, ...)
 {
 	va_list ap;
+	struct einfo_term *term;
+	ECOLOR color;
+	const char *msg;
 
-	if (is_quiet())
+	if (!is_verbose()) {
 		return retval;
+	}
+
 	va_start(ap, fmt);
-	_do_eend("ewend", retval, fmt, ap);
+
+	if (retval == 0) {
+		term = &stdout_term;
+		color = ECOLOR_GOOD;
+		msg = "ok";
+	} else {
+		term = &stderr_term;
+		color = ECOLOR_BAD;
+		msg = "!!";
+	}
+	_eend_message(term, retval, color, fmt, ap);
+	_eend_status(term, -1, color, msg);
+
 	va_end(ap);
-	LASTCMD("ewend");
+	LASTCMD(__func__);
+	return retval;
+}
+int ewendv(int retval, const char * EINFO_RESTRICT fmt, ...)
+{
+	va_list ap;
+	struct einfo_term *term;
+	ECOLOR color;
+	const char *msg;
+
+	if (!is_verbose()) {
+		return retval;
+	}
+
+	va_start(ap, fmt);
+
+	if (retval == 0) {
+		term = &stdout_term;
+		color = ECOLOR_GOOD;
+		msg = "ok";
+	} else {
+		term = &stderr_term;
+		color = ECOLOR_WARN;
+		msg = "!!";
+	}
+	_eend_message(term, retval, color, fmt, ap);
+
+	/* Do not use the warn color for "!!" */
+	if (retval != 0) {
+		color = ECOLOR_BAD;
+	}
+	_eend_status(term, -1, color, msg);
+
+	va_end(ap);
+	LASTCMD(__func__);
 	return retval;
 }
 
-void
-ebracket(int col, ECOLOR color, const char *msg)
+/*
+ * TODO:
+ * 	- proper test progs
+ */
+void eindent(void)
 {
-	_eend(stdout, col, color, msg);
-}
+	size_t new_level = indent_level + 1;
 
-void
-eindent(void)
-{
-	char *env = getenv("EINFO_INDENT");
-	int amount = 0;
-	char *num;
-
-	if (env) {
-		errno = 0;
-		amount = strtoimax(env, NULL, 0);
-		if (errno != 0)
-			amount = 0;
+	if (new_level * INDENT_WIDTH <= INDENT_MAX) {
+		indent_level = new_level;
 	}
-	amount += INDENT_WIDTH;
-	if (amount > INDENT_MAX)
-		amount = INDENT_MAX;
-	xasprintf(&num, "%08d", amount);
-	setenv("EINFO_INDENT", num, 1);
-	free(num);
 }
-
 void eoutdent(void)
 {
-	char *env = getenv("EINFO_INDENT");
-	int amount = 0;
-	char *num = NULL;
-	int serrno = errno;
-
-	if (!env)
-		return;
-	errno = 0;
-	amount = strtoimax(env, NULL, 0);
-	if (errno != 0)
-		amount = 0;
-	else
-		amount -= INDENT_WIDTH;
-	if (amount <= 0)
-		unsetenv("EINFO_INDENT");
-	else {
-		xasprintf(&num, "%08d", amount);
-		setenv("EINFO_INDENT", num, 1);
-		free(num);
+	if (indent_level > 0) {
+		indent_level--;
 	}
-	errno = serrno;
 }
-
-int
-einfovn(const char *EINFO_RESTRICT fmt, ...)
+void eindentv(void)
 {
-	int retval;
-	va_list ap;
-
-	if (!fmt || !is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
-	va_end(ap);
-	LASTCMD("einfovn");
-	return retval;
-}
-
-int
-ewarnvn(const char *EINFO_RESTRICT fmt, ...)
-{
-	int retval;
-	va_list ap;
-
-	if (!fmt || !is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	retval = _ewarnvn(fmt, ap);
-	va_end(ap);
-	LASTCMD("ewarnvn");
-	return retval;
-}
-
-int
-einfov(const char *EINFO_RESTRICT fmt, ...)
-{
-	int retval;
-	va_list ap;
-
-	if (!fmt || !is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
-	retval += printf("\n");
-	va_end(ap);
-	LASTCMD("einfov");
-	return retval;
-}
-
-int
-ewarnv(const char *EINFO_RESTRICT fmt, ...)
-{
-	int retval;
-	va_list ap;
-
-	if (!fmt || !is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	retval = _ewarnvn(fmt, ap);
-	retval += printf("\n");
-	va_end(ap);
-	LASTCMD("ewarnv");
-	return retval;
-}
-
-int
-ebeginv(const char *EINFO_RESTRICT fmt, ...)
-{
-	int retval;
-	va_list ap;
-
-	if (!fmt || !is_verbose())
-		return 0;
-
-	va_start(ap, fmt);
-	retval = _einfovn(fmt, ap);
-	retval += printf(" ...");
-	if (colour_terminal(stdout))
-		retval += printf("\n");
-	va_end(ap);
-	LASTCMD("ebeginv");
-	return retval;
-}
-
-int
-eendv(int retval, const char *EINFO_RESTRICT fmt, ...)
-{
-	va_list ap;
-
-	if (!is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	_do_eend("eendv", retval, fmt, ap);
-	va_end(ap);
-	LASTCMD("eendv");
-	return retval;
-}
-
-int
-ewendv(int retval, const char *EINFO_RESTRICT fmt, ...)
-{
-	va_list ap;
-
-	if (!is_verbose())
-		return 0;
-	va_start(ap, fmt);
-	_do_eend("ewendv", retval, fmt, ap);
-	va_end(ap);
-	LASTCMD("ewendv");
-	return retval;
-}
-
-void
-eindentv(void)
-{
-	if (is_verbose())
+	if (is_verbose()) {
 		eindent();
+	}
+}
+void eoutdentv(void)
+{
+	if (is_verbose()) {
+		eoutdent();
+	}
 }
 
-void
-eoutdentv(void)
+void eprefix(const char * EINFO_RESTRICT p)
 {
-	if (is_verbose())
-		eoutdent();
+	prefix = p;
 }
 
 /* TODO: remove */
